Remove the 'Welcome' step during two-factor setup.
The 'Welcome' step is removed since it only informs the user they
are about to enable two-factor authentication, and requires the
user to click 'Next' to proceed.

This step is not needed since if the user is on the setup page
already, the user is aware it is to set up two-factor auth.